maturitylevel/requirements-dev.txt
-r requirements.txt
pytest
pytest-django
pytest-mock
requests